requirements.txt
discord.py
google-generativeai
PyYAML
Flask
python-dotenv
watchdog
waitress